pyproject.toml
[project]
name = "advance-mcp-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py>=2.16.3",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
